requirements.txt
flask
flask-sqlalchemy
mistralai
werkzeug
PyPDF2
python-docx
gunicorn
psycopg2-binary
